% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_contour.R
\name{active_contour}
\alias{active_contour}
\title{Evolves an active contour}
\usage{
active_contour(phi, g, n_iter = 2000, v = 1, dt = 1, buffer = 3,
  show = FALSE, orig_image = NA)
}
\arguments{
\item{phi}{an inital contour}

\item{g}{the stopping function of the contour}

\item{n_iter}{the maximal number of iterations}

\item{v}{the velocity of the development of the contour}

\item{dt}{the time interval of the contour development}

\item{buffer}{the buffer of the contour around the shape}

\item{show}{should the development of the contour be shown interactively}

\item{orig_image}{the original image on which the contour should be developed. Only necessary if \code{show=TRUE}}
}
\value{
the active contour around the features of the image
}
\description{
Evolves an active contour
}
