% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing_utils.R
\name{add_canvas}
\alias{add_canvas}
\title{Add canvas to an image}
\usage{
add_canvas(x, nrows, ncols, center = T, col = "white")
}
\arguments{
\item{x}{the original image, either of class \code{EBImage::Image} or \code{matrix}.}

\item{nrows, ncols}{The number of rows and columns of which the image should be enlarged}

\item{center}{should the original picture be centered within the canvas?}

\item{col}{the color of the canvas}
}
\value{
The image with an enlarged canvas in
}
\description{
Add canvas to an image of class \code{EBImage}.
}
