% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_extraction.R
\name{contour_vector_fg}
\alias{contour_vector_fg}
\title{Get the profile contour by the first foreground pixel}
\usage{
contour_vector_fg(x)
}
\arguments{
\item{x}{the image of the profile of the vessel}
}
\value{
a vector with the position of the profile
}
\description{
Get the profile contour by the first foreground pixel
}
