% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel_cropping.R
\name{crop_top}
\alias{crop_top}
\title{Crop the image of a ceramic vessels to its top corner points}
\usage{
crop_top(x)
}
\arguments{
\item{x}{the image of a vessel}
}
\value{
the vessel image cropped
}
\description{
Crop the image of a ceramic vessels to its top corner points
}
