% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_contour_utils.R
\name{curvature}
\alias{curvature}
\title{Calculate the curvature of a active contour phi}
\usage{
curvature(f)
}
\arguments{
\item{f}{the active contour function phi}
}
\value{
the curvature of that function
}
\description{
Calculate the curvature of a active contour phi
}
