% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_contour_utils.R
\name{default_phi}
\alias{default_phi}
\title{Return the initial phi surface for active contour}
\usage{
default_phi(x)
}
\arguments{
\item{x}{a matrix containing the values of the image}
}
\value{
a matrix of the same dimensions
}
\description{
The function returns the initialize contour surface phi,
5px from the image border. Values outside the curve are set to 1,
inside the curve to -1.
}
