% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_contour_utils.R
\name{div}
\alias{div}
\title{Calculate the divergence of two gradient derivatives}
\usage{
div(fx, fy)
}
\arguments{
\item{fx, fy}{the x and y gradient derivatives}
}
\value{
the divergence of those gradients
}
\description{
Calculate the divergence of two gradient derivatives
}
