% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_contour_utils.R
\name{dot}
\alias{dot}
\title{Sum the Product of list of matrixes}
\usage{
dot(x, y)
}
\arguments{
\item{x, y}{two lists of matrices}
}
\value{
the sum of the products
}
\description{
Sum the Product of list of matrixes
}
