% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel_cropping.R
\name{get_bottom_points}
\alias{get_bottom_points}
\title{Get the bottom break points of a ceramic vessels}
\usage{
get_bottom_points(x)
}
\arguments{
\item{x}{the image of a vessel}
}
\value{
the two lowest break point of the vessel bottom
}
\description{
Get the bottom break points of a ceramic vessels
}
