% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel_cropping.R
\name{get_corner}
\alias{get_corner}
\title{Returns the break point of a vessel contour from one of its quandrants}
\usage{
get_corner(x)
}
\arguments{
\item{x}{the image of a vessel quadrant}
}
\value{
the break point of that vessel quadrant
}
\description{
Returns the break point of a vessel contour from one of its quandrants
}
