% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel_cropping.R
\name{get_top_points}
\alias{get_top_points}
\title{Get the top break points of a ceramic vessels}
\usage{
get_top_points(x)
}
\arguments{
\item{x}{the image of a vessel}
}
\value{
the two break point of the vessel top
}
\description{
Get the top break points of a ceramic vessels
}
