% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing_utils.R
\name{getMinBBox}
\alias{getMinBBox}
\title{Minimum-Area Bounding Box For A Set Of 2D-Points}
\usage{
getMinBBox(xy)
}
\arguments{
\item{xy}{either a numerical (n x 2)-matrix with the (x,y)-coordinates of n >= 2 points (1 row of coordinates per point), or a data frame with either the variables \code{x}, \code{y} or \code{point.x}, \code{point.y}.}
}
\value{
A list with the following information about the minimum-area bounding box:
\item{pts}{a (4 x 2)-matrix containing the coordinates of the (ordered) vertices.}
\item{width}{width of the box.}
\item{height}{height of the box.}
\item{FoM}{figure of merit, i.e., the average side length of the box: (\code{width} + \code{height}) / 2.}
\item{diag}{length of box diagonal.}
\item{angle}{orientation of the box' longer edge pointing up as returned by \code{\link{atan2}}, but in degree.}
}
\description{
Calculates the vertices of the minimum-area, possibly oriented bounding box given a set of 2D-coordinates. Adapted from shotGroups package.
}
