% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_contour_utils.R
\name{grad}
\alias{grad}
\title{Calculate the central difference (derivative) gradients (x,y) of a (image) matrix}
\usage{
grad(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a list of two derivative matrices
}
\description{
Calculate the central difference (derivative) gradients (x,y) of a (image) matrix
}
\examples{
my_matrix <- matrix(c(0,1,1,0), nrow=2)
grad(my_matrix)
}
