% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad1_c}
\alias{grad1_c}
\title{A simple script to get the central difference gradient of a (image) matrix}
\usage{
grad1_c(x)
}
\arguments{
\item{x}{a matrix representing the image}
}
\value{
the first derivative (central difference gradient) of that matrix
}
\description{
A simple script to get the central difference gradient of a (image) matrix
}
