% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing_utils.R
\name{img_crop_backgroud}
\alias{img_crop_backgroud}
\title{Crop the image to content}
\usage{
img_crop_backgroud(x, bg = 0)
}
\arguments{
\item{x}{the image}

\item{bg}{the background value}
}
\value{
the cropped image
}
\description{
Crop the an image of class \code{EBImage} to only those pixels that are not background
}
