% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_contour_utils.R
\name{norm}
\alias{norm}
\title{Calculate the euclidian norm of two matrices}
\usage{
norm(x)
}
\arguments{
\item{x}{a list of two matrices}
}
\value{
the norm of those matrices
}
\description{
Calculate the euclidian norm of two matrices
}
\examples{
my_matrix <- matrix(c(0,1,1,0), nrow=2)
norm(grad(my_matrix))
}
