% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_extraction.R
\name{profile_to_image}
\alias{profile_to_image}
\title{Turns a contour into an image again}
\usage{
profile_to_image(x, columns_out)
}
\arguments{
\item{x}{the contour vector}

\item{columns_out}{the width of the resulting image}
}
\value{
an image representing the profile contour
}
\description{
Turns a contour into an image again
}
