% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{shapAAR}
\alias{shapAAR}
\alias{shapAAR-package}
\title{shapAAR}
\description{
Identification, processing, analysis and classification of
shapes (not only) for archaeological objects
}
\author{
Martin Hinz
}
