% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapaar_bulk_convert.R
\name{shapaar_bulk_convert}
\alias{shapaar_bulk_convert}
\title{Extracting the conture from a bulk of scanned images in one folder}
\usage{
shapaar_bulk_convert(p, ext = "png", thresh = 0.02, v = 1, dt = 1,
  n_iter = 1000, buffer = 2, vectorlength = 100,
  side_of_profile = c("both", "left", "right"))
}
\arguments{
\item{p}{the path to the directory with the images (results will be placed alongside originals)}

\item{ext}{the file extention to be used; at the moment only PNG is supported)}

\item{thresh}{threshold value for the comparison of left and right side}

\item{v}{the velocity of the active contour (default = 1)}

\item{dt}{the time interval that is used for the development of the contour (default = 1)}

\item{n_iter}{maximum iteration length of the algorithm (default = 1000)}

\item{buffer}{buffer of the contour (default = 2)}

\item{vectorlength}{desired length of the resulting contour vector (default = 100)}

\item{side_of_profile}{side of the drawing to be used as contour ('both' (mean of l+r), 'left', 'right' defaulting to 'left')}
}
\description{
Extracting the conture from a bulk of scanned images in one folder
}
