% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_contour_utils.R
\name{simple_edge_detection}
\alias{simple_edge_detection}
\title{A simple edge detection using the image derivatives}
\usage{
simple_edge_detection(x)
}
\arguments{
\item{x}{a matrix containing the intensity values (grayscale)}
}
\value{
a matrix containing the edges
}
\description{
A simple edge detection using the image derivatives
}
