% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_contour_utils.R
\name{stopping_fun}
\alias{stopping_fun}
\title{Stopping function for the active contour algorithm}
\usage{
stopping_fun(x)
}
\arguments{
\item{x}{a matrix containing the intensity values (grayscale)}
}
\value{
the stopping function
}
\description{
The stopping function for the active contour algorithm,
in this implementation using the edges of the image.
}
